add path aliases for cleaner imports
