+++ src/main.js
@@ -1,7 +1,7 @@
 import './style.scss'
-import { renderListIntoContainer } from './components/Card'
-import { items } from './constants'
+import { renderProductCardListIntoContainer } from 'components/ProductCardList'
+import { items } from 'utils/constants'
 
 document.addEventListener('DOMContentLoaded', function () {
-  renderListIntoContainer('#root', items)
+  renderProductCardListIntoContainer('#root', items)
 })
