+++ src/utils/throttle.js
@@ -1,0 +1,13 @@
+let throttleTimer
+const throttle = (callback, time) => {
+  if (throttleTimer) return
+
+  throttleTimer = true
+
+  setTimeout(() => {
+    callback()
+    throttleTimer = false
+  }, time)
+}
+
+export { throttle }
