+++ webpack.common.js
@@ -27,4 +27,10 @@ module.exports = {
       },
     ],
   },
+  resolve: {
+    alias: {
+      components: path.resolve(__dirname, 'src/components/'),
+      utils: path.resolve(__dirname, 'src/utils/'),
+    },
+  },
 }
